4.22 (completed all questions save for 4.6, 4.7, and 4.23)

+++ controllers/blogs.js
@@ -1,15 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-	const authorization = request.get('authorization')
-	if (authorization && authorization.startsWith('Bearer ')) {
-	  return authorization.replace('Bearer ', '')
-	}
-	return null
-}
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
 	const blogs = await Blog
@@ -26,13 +17,9 @@ blogsRouter.get('/:id', async (request, response) => {
 	}
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', [middleware.userExtractor], async (request, response) => {
 	const body = request.body
-	const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  	if (!decodedToken.id) {
-   		return response.status(401).json({ error: 'token invalid' })
-  	}
-  	const user = await User.findById(decodedToken.id)
+	const user = request.user
 
 	const blog = new Blog({
 		title: body.title,
@@ -49,7 +36,14 @@ blogsRouter.post('/', async (request, response) => {
 	response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', [middleware.userExtractor], async (request, response) => {
+	const user = request.user
+	const blog = await user.blogs.find(b => b.toString() === request.params.id)
+
+	if (!blog) {
+		return response.status(401).json({ error: 'only the blog creator can delete the blog' })
+	}
+	
 	await Blog.findByIdAndRemove(request.params.id)
 	response.status(204).end()
 })
